+++ R/zzz.R
@@ -1,1 +1,1 @@
-utils::globalVariables(c("CI", "fit", "geom_hline", "se.fit"))
+utils::globalVariables(c("CI", "fit", "geom_hline", "se.fit", "SE", "CI_lower", "CI_upper", ".idx"))