+++ NEWS.md
@@ -1,8 +1,11 @@
 # Change Log
 
 ## [Unreleased] - 2020-12-20
+### Added
+- reintroduced `plot_difference()`.
+
 ### Changed
-- `get_gam_predictions()` now returns an `.idx` column which assigns a number to each curve (when multiple variables are used). The `.idx` column can be used to correctly group `geom_ribbon()` for plotting CIs. 
+- `get_gam_predictions()` now returns an `.idx` column which assigns a number to each curve (when multiple variables are used). The `.idx` column can be used to correctly group `geom_ribbon()` for plotting CIs.
 
 ## 3.0.0 - 2020-04-24
 ### Removed