+++ NEWS.md
@@ -2,10 +2,13 @@
 
 ## Added
 
-* ✨ `get_smooths_difference()` which returns a tibble with the difference of two levels of a smooth (#11).
+* ✨ - `get_smooths_difference()` which returns a tibble with the difference of two levels of a smooth (#11).
 
-* 🍇 `plot-smooths.Rmd` now includes a plotting example with `get_smooths_difference()`.
+* 🍇 - `plot-smooths.Rmd` now includes a plotting example with `get_smooths_difference()`.
 
+## Changed
+
+* ⬆️ - Use tidy evaluation.
 
 # tidymv v3.1.0